agentic-mirror/backend/requirements.txt
# Agentic Mirror — Backend Dependencies
# See AGENT.md §4 — Tech Stack (Backend)
# See AGENT.md §13 — Environment Setup

fastapi
uvicorn[standard]
anthropic
openai
scikit-learn
umap-learn
python-dotenv
numpy
pydantic
